requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
pgvector==0.3.6
redis==5.2.1
rq==2.0.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
httpx==0.28.1
anthropic==0.40.0
reportlab==4.2.5
gitpython==3.1.44
boto3==1.35.88
fastembed==0.4.2
